requirements.txt
CouchDB==1.1
Django==1.10.5
django-filter==1.0.1
djangorestframework==3.5.3
Markdown==2.6.7
pkg-resources==0.0.0
